MAR-46 move root level dynamic route into static route
