+++ src/app/docs/[documentation]/(categories)/[category]/[slug]/page.tsx
@@ -1,4 +1,4 @@
-import Doc from '@/app/[documentation]/(categories)/templates/doc';
+import Doc from '@/app/docs/[documentation]/(categories)/templates/doc';
 import { DynamicRoute, PostData } from '@/app/types';
 import { GetDataContent } from '@/utils/mdx';
 import { Metadata } from 'next';
