+++ src/app/docs/[documentation]/(categories)/[category]/page.tsx
@@ -1,6 +1,6 @@
 import type { GlobalButtonSettings } from '@/app/components/shared/button';
 import GlobalCard from '@/app/components/shared/card';
-import Doc from '@/app/[documentation]/(categories)/templates/doc';
+import Doc from '@/app/docs/[documentation]/(categories)/templates/doc';
 import { DynamicRoute, PostData, BasicPostData } from '@/app/types';
 import { GetDataContent, GetSubFolders } from '@/utils/mdx';
 import { Metadata } from 'next';
